server/rag_server/requirements.txt
chromadb==0.4.16

langchain>=0.3.0
langchain-core
langchain-chroma

langchain_ollama
langchain_community

langchain-google-genai
langchain-text-splitters
pypdf
fastapi
uvicorn[standard]
pydantic==2.*
onnxruntime==1.17.3
python-dotenv

pandas
numpy<2
cryptography
